backend/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# automatic_czech_ass/main.py: 3
CherryPy == 18.6.0
cherrypy-cors == 1.6


# automatic_czech_ass/main.py: 2
Jinja2 == 2.11.3

# automatic_czech_ass/formats/ass.py: 3,4
ass-tag-parser == 2.3.1

# automatic_czech_ass/main.py: 9
psycopg2 == 2.8.6

# automatic_czech_ass/formats/ass.py: 1
# automatic_czech_ass/test.py: 1
pysubs2 == 1.0.0

# automatic_czech_ass/main.py: 5
# automatic_czech_ass/translator.py: 1
requests == 2.25.1

python-dotenv == 0.17.0
